+++ packages/@leight/file-client/src/sdk/ClientSourceProvider.tsx
@@ -59,4 +59,4 @@ export const FileQueryProvider: FC<IFileQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ubs2hujmp5erwcdrf47jek9e               = true;
+export const $leight_vf7isc0yzpi1ianungdrcd48               = true;