+++ packages/@leight/file-client/src/sdk/ClientStore.ts
@@ -43,4 +43,4 @@ export const FileSortStore                    = createSortContext<IFileSortSchem
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_vfdqh0dbvc3n059uszvszk5a = true;
+export const $leight_skzznflalrhspi6kid9qt24j = true;