+++ packages/@leight/file/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const FileQuerySchema                  = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_fm1edk79pywq0h5so18wiusm = true;
+export const $leight_c3evffqoxwnka31e8w3gjkcr = true;