+++ packages/@leight/job-client/src/sdk/ClientSourceProvider.tsx
@@ -59,4 +59,4 @@ export const JobQueryProvider: FC<IJobQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_xv8rftnwicqm4f0w1sr0oolf             = true;
+export const $leight_pd6ux7assjwvc4jdsjjg7mau             = true;