+++ packages/@leight/job-client/src/sdk/ClientStore.ts
@@ -43,4 +43,4 @@ export const JobSortStore                     = createSortContext<IJobSortSchema
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_srdb93wgbpwoyyhu66wj8exo = true;
+export const $leight_lspkm3s1346kkjizadn7cca5 = true;