+++ packages/@leight/job-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const JobSourceProcedure               = withSourceProcedure<IJobSourceSc
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_k2v56v96arf946rjs0ohcq9v = true;
+export const $leight_thyqziubc01lbml3z4qyb6ys = true;