+++ packages/@leight/job/src/sdk/PrismaSchema.ts
@@ -72,4 +72,4 @@ export const JobQuerySchema                   = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_wkawvw0pqzij3x19llpq6608 = true;
+export const $leight_blx778bzst0jpyor12dgi2h7 = true;