+++ packages/@leight/job/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const JobSourceContext                 = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_unlikpedjojw1jdc9b1o6rjg = true;
+export const $leight_kl0d66b9rgqtdqqa0eduxntz = true;