+++ packages/@leight/user-server/src/sdk/ServerPrismaSource.ts
@@ -73,7 +73,7 @@ export class UserBasePrismaSource extends AbstractSource<IUserSourceSchema> {
     }
 
     toWhere(filter?: IUserSourceSchema["Filter"]): IUserWhere | undefined {
-        return undefined;
+        return filter;
     }
 
     toWhereUnique(filter?: IUserSourceSchema["Filter"]): IUserWhereUnique | undefined {
@@ -89,4 +89,4 @@ export class UserBasePrismaSource extends AbstractSource<IUserSourceSchema> {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_si715m5e7sihbw8qlzw5p21j = true;
+export const $leight_warlnb5bulfveen8k15j5bt1 = true;