+++ packages/@leight/user-server/src/sdk/ServerSource.ts
@@ -13,4 +13,4 @@ export class UserSource extends UserSourceEx implements IUserSource {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_kd415xs09iq1j83wt8kb3wyb = true;
+export const $leight_xtkh15t1x46fr8ckwqmsy06n = true;