+++ packages/@leight/user/src/sdk/Source.ts
@@ -20,4 +20,4 @@ export const UserSourceContext                = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_hilav776ztlisr38tfgt7dw2 = true;
+export const $leight_jc2txikoova8uv01p62lv1l3 = true;