+++ packages/@leight/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/sdk
 
+## 0.5.21
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.20
 
 ### Patch Changes