+++ packages/@leight/user/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/user
 
+## 0.5.74
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.80
+
 ## 0.5.73
 
 ### Patch Changes